requirements.txt
bcrypt==4.3.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.3
idna==3.10
paramiko==3.5.1
psycopg2-binary==2.9.10
pycparser==2.22
PyNaCl==1.5.0
python-dotenv==1.1.0
requests==2.32.3
sshtunnel==0.4.0
urllib3==2.4.0
